membuat class EmployeeDependent
