+++ src/lib/EmployeeDependent.java
@@ -1,0 +1,29 @@
+package lib;
+
+import java.util.List;
+import java.util.LinkedList;
+
+public class EmployeeDependent {
+    private String spouseName;
+    private String spouseIdNumber;
+
+    private List<String> childNames;
+    private List<String> childIdNumbers;
+
+    public EmployeeDependent() {
+        this.spouseName = "";
+        this.spouseIdNumber = "";
+        this.childNames = new LinkedList<>();
+        this.childIdNumbers = new LinkedList<>();
+    }
+
+    public void setSpouse(String spouseName, String spouseIdNumber) {
+        this.spouseName = spouseName;
+        this.spouseIdNumber = spouseIdNumber;
+    }
+
+    public void addChild(String childName, String childIdNumber) {
+        childNames.add(childName);
+        childIdNumbers.add(childIdNumber);
+    }
+}
